Renovate storymap
